oh-tab-ycg2: extract sanitizeChatMessages
